requirements.txt
dbt-postgres
psycopg2-binary
kafka-python
Faker
prometheus-client
Flask
jwt
